scripts/pyproject.toml
[tool.poetry]
name = "portfolio-updater"
version = "0.1.0"
description = "Fetches asset prices and exchange rates, then writes them to Google Sheets"
authors = []
readme = "README.md"
packages = [{ include = "portfolio_updater", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
gspread = "^6.1"
google-auth = "^2.29"
yfinance = "^0.2"
pydantic = ">=2.0"
pydantic-settings = "^2.3"
loguru = "^0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"

[tool.poetry.scripts]
portfolio-updater = "portfolio_updater.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
